Replace platform specific read/write implementation with FileStream usage
