+++ library/file/src/androidNativeMain/kotlin/io/matthewnelson/kmp/file/internal/AndroidNativePlatform.kt
@@ -56,10 +56,10 @@ private val __TempDir: File by lazy(LazyThreadSafetyMode.SYNCHRONIZED) {
 
 @Throws(Exception::class)
 @OptIn(DelicateFileApi::class, ExperimentalForeignApi::class)
-private fun readPackageName(pid: String): String = "/proc/$pid/cmdline".toFile().openR().use { file ->
+private fun readPackageName(pid: String): String = "/proc/$pid/cmdline".toFile().openRead().use { s ->
     val buf = ByteArray(512)
-    val read = file.fRead(buf)
-    if (read == -1) throw errnoToIOException(errno)
+    val read = s.read(buf)
+    check(read >= 3) { "read < 3" }
 
     // Strip off trailing null bytes (0)
     val iZero = buf.indexOfFirst { it == 0.toByte() }
