+++ library/file/src/commonMain/kotlin/io/matthewnelson/kmp/file/FileStreams.kt
@@ -158,64 +158,3 @@ public expect sealed interface FileStream: Closeable {
 
     // TODO: ReadWrite: Read, Write
 }
-
-// Strictly for supporting isInstance checks
-internal class FileStreamReadOnly internal constructor(private val s: AbstractFileStream): FileStream.Read by s {
-    init {
-        check(s.canRead) { "AbstractFileStream.canRead != true" }
-        check(!s.canWrite) { "AbstractFileStream.canWrite != false" }
-    }
-    override fun position(new: Long): FileStream.Read { s.position(new); return this }
-    override fun equals(other: Any?): Boolean = other is FileStreamReadOnly && other.s == s
-    override fun hashCode(): Int = s.hashCode()
-    override fun toString(): String = s.toString()
-}
-
-// Strictly for supporting isInstance checks
-internal class FileStreamWriteOnly internal constructor(private val s: AbstractFileStream): FileStream.Write by s {
-    init {
-        check(!s.canRead) { "AbstractFileStream.canRead != false" }
-        check(s.canWrite) { "AbstractFileStream.canWrite != true" }
-    }
-    override fun equals(other: Any?): Boolean = other is FileStreamWriteOnly && other.s == s
-    override fun hashCode(): Int = s.hashCode()
-    override fun toString(): String = s.toString()
-}
-
-internal abstract class AbstractFileStream internal constructor(
-    internal val canRead: Boolean,
-    internal val canWrite: Boolean,
-): FileStream.Read, FileStream.Write {
-
-    init {
-        if (!canRead && !canWrite) throw IllegalStateException("!canRead && !canWrite")
-    }
-
-    final override fun read(buf: ByteArray): Int = read(buf, 0, buf.size)
-    final override fun write(buf: ByteArray) { write(buf, 0, buf.size) }
-
-    // Read
-    override fun position(): Long {
-        throw IllegalStateException("FileStream is O_WRONLY")
-    }
-    override fun position(new: Long): FileStream.Read {
-        throw IllegalStateException("FileStream is O_WRONLY")
-    }
-    override fun read(buf: ByteArray, offset: Int, len: Int): Int {
-        throw IllegalStateException("FileStream is O_WRONLY")
-    }
-    override fun size(): Long {
-        throw IllegalStateException("FileStream is O_WRONLY")
-    }
-
-    // Write
-    override fun flush() {
-        throw IllegalStateException("FileStream is O_RDONLY")
-    }
-    override fun write(buf: ByteArray, offset: Int, len: Int) {
-        throw IllegalStateException("FileStream is O_RDONLY")
-    }
-}
-
-@Suppress("NOTHING_TO_INLINE")
-internal inline fun fileStreamClosed(): IOException = IOException("FileStream is closed")
