+++ library/file/src/commonTest/kotlin/io/matthewnelson/kmp/file/ReadUnitTest.kt
@@ -16,17 +16,21 @@
 package io.matthewnelson.kmp.file
 
 import io.matthewnelson.kmp.file.internal.IsWindows
+import io.matthewnelson.kmp.file.internal.commonReadBytes
+import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
+import kotlin.test.assertFalse
 import kotlin.test.assertTrue
+import kotlin.test.fail
 
 class ReadUnitTest: ReadSharedTest() {
 
     @Test
     fun givenFile_whenReadBytes_thenSha256MatchesExpected() = skipTestIf(isJsBrowser || IS_ANDROID) {
         val expected = if (IsWindows) {
-            // Windows will produce a different result because of its
-            // EOL value.
+            // Windows will produce a different result because of its EOL value.
             "9d8f1dc39b0fc445f5c85b23dc6cdcb156bb166e0974c62ee5ffa82b590d417c"
         } else {
             "439664467fd3b26829244d7bb87b20e7873a97e494c6ead836d359d90254b76f"
@@ -41,5 +45,68 @@ class ReadUnitTest: ReadSharedTest() {
     @Test
     fun givenFile_whenReadUtf8_thenSomethingIsReturned() = skipTestIf(isJsBrowser || IS_ANDROID) {
         assertTrue(FILE_LOREM_IPSUM.readUtf8().isNotBlank())
+    }
+
+    @Test
+    fun givenFile_whenReadStreamSizeReportsInaccurately_thenReadBytesReturnsExpected() = skipTestIf(isJsBrowser) {
+        val tmp = randomTemp()
+        val expected = Random.Default.nextBytes(212_121)
+        tmp.writeBytes(expected)
+
+        try {
+            // Simulates the filesystem lying to us about what the actual size
+            // of a file is to ensure that it is actually being read fully until
+            // the stream reaches exhaustion (i.e. read returns -1).
+            longArrayOf(
+                0L,
+                10L,
+                (expected.size - 1).toLong(),
+                (expected.size + 1).toLong(),
+            ).forEach { testSize ->
+                val stream = object : TestStream(s = tmp.openRead()) {
+                    override fun size(): Long = testSize
+                }
+                val actual = tmp.commonReadBytes(open = { stream })
+
+                assertFalse(stream.isOpen())
+                assertEquals(expected.size, actual.size)
+                for (i in expected.indices) {
+                    val e = expected[i]
+                    val a = actual[i]
+                    assertEquals(e, a, "expected[$e] != actual[$a] >> index[$i] - testSize[$testSize]")
+                }
+            }
+        } finally {
+            tmp.delete2()
+        }
+    }
+
+    @Test
+    fun givenFile_whenReadStreamSizeReturnsGreaterThanIntMax_thenThrowsFileSystemException() = skipTestIf(isJsBrowser) {
+        val tmp = randomTemp()
+        tmp.writeUtf8("Hello World!")
+        var stream: TestStream? = null
+
+        try {
+            stream = object : TestStream(s = tmp.openRead()) {
+                override fun size(): Long = Int.MAX_VALUE.toLong() + 1L
+            }
+            tmp.commonReadBytes { stream }
+            fail("readBytes should have thrown exception...")
+        } catch (e: FileSystemException) {
+            assertEquals(true, e.message?.contains("Size exceeds maximum"))
+            assertEquals(false, stream?.isOpen())
+        } finally {
+            tmp.delete2()
+        }
+    }
+
+    private abstract class TestStream(val s: FileStream.Read): AbstractFileStream(canRead = true, canWrite = false) {
+        final override fun isOpen(): Boolean = s.isOpen()
+        final override fun position(): Long = s.position()
+        final override fun position(new: Long): FileStream.Read { s.position(new); return this }
+        final override fun read(buf: ByteArray, offset: Int, len: Int): Int = s.read(buf, offset, len)
+        override fun size(): Long = s.size()
+        final override fun close() { s.close() }
     }
 }
