+++ library/file/src/jsMain/kotlin/io/matthewnelson/kmp/file/FileJs.kt
@@ -15,8 +15,9 @@
  **/
 package io.matthewnelson.kmp.file
 
-import io.matthewnelson.kmp.file.internal.*
+import io.matthewnelson.kmp.file.internal.fileNotFoundException
 import io.matthewnelson.kmp.file.internal.fs.FsJsNode
+import io.matthewnelson.kmp.file.internal.require
 
 /**
  * Retrieve the symbolic [Stats] referred to by the abstract pathname.
