+++ library/file/src/jsMain/kotlin/io/matthewnelson/kmp/file/internal/fs/FsJsNode.kt
@@ -28,7 +28,7 @@ import io.matthewnelson.kmp.file.IOException
 import io.matthewnelson.kmp.file.NotDirectoryException
 import io.matthewnelson.kmp.file.OpenExcl
 import io.matthewnelson.kmp.file.errorCodeOrNull
-import io.matthewnelson.kmp.file.fileStreamClosed
+import io.matthewnelson.kmp.file.internal.fileStreamClosed
 import io.matthewnelson.kmp.file.internal.Mode
 import io.matthewnelson.kmp.file.internal.Path
 import io.matthewnelson.kmp.file.internal.checkBounds
