+++ library/file/src/jvmMain/kotlin/io/matthewnelson/kmp/file/internal/-JvmException.kt
@@ -1,0 +1,39 @@
+/*
+ * Copyright (c) 2025 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+@file:Suppress("NOTHING_TO_INLINE")
+
+package io.matthewnelson.kmp.file.internal
+
+import io.matthewnelson.kmp.file.File
+import io.matthewnelson.kmp.file.AccessDeniedException
+import io.matthewnelson.kmp.file.FileAlreadyExistsException
+import io.matthewnelson.kmp.file.FileSystemException
+
+internal inline fun SecurityException.toAccessDeniedException(file: File): AccessDeniedException {
+    return AccessDeniedException(file, reason = "SecurityException[${message}]")
+}
+
+internal actual inline fun fileAlreadyExistsException(
+    file: File,
+    other: File?,
+    reason: String?,
+): FileAlreadyExistsException = FileAlreadyExistsException(file, other, reason)
+
+internal actual inline fun fileSystemException(
+    file: File,
+    other: File?,
+    reason: String?,
+): FileSystemException = FileSystemException(file, other, reason)
