+++ library/file/src/jvmMain/kotlin/io/matthewnelson/kmp/file/internal/fs/-FsJvmAndroid.kt
@@ -28,7 +28,7 @@ import io.matthewnelson.kmp.file.FsInfo
 import io.matthewnelson.kmp.file.IOException
 import io.matthewnelson.kmp.file.NotDirectoryException
 import io.matthewnelson.kmp.file.OpenExcl
-import io.matthewnelson.kmp.file.fileStreamClosed
+import io.matthewnelson.kmp.file.internal.fileStreamClosed
 import io.matthewnelson.kmp.file.internal.Mode
 import io.matthewnelson.kmp.file.internal.Mode.Mask.Companion.convert
 import io.matthewnelson.kmp.file.internal.fileNotFoundException
