+++ library/file/src/jvmTest/kotlin/io/matthewnelson/kmp/file/ChmodFsJvmDefaultUnitTest.kt
@@ -15,7 +15,7 @@
  **/
 package io.matthewnelson.kmp.file
 
-import io.matthewnelson.kmp.file.internal.commonChmod
+import io.matthewnelson.kmp.file.internal.fs.commonChmod
 import io.matthewnelson.kmp.file.internal.fs.FsJvmDefault
 import kotlin.test.Ignore
 import kotlin.test.Test
